requirements.txt
PyJWT~=1.6.4
python-dotenv~=0.8.2
tornado~=5.0.2
tornado-sqlalchemy~=0.4.1
SQLAlchemy~=1.2.8
werkzeug~=0.14.1
wtforms~=2.2.1
wtforms-tornado~=0.0.2